Move OverrideNodeConfig to its own file
