+++ config/src/config/mod.rs
@@ -21,6 +21,7 @@ mod netbench;
 mod network_config;
 mod node_config;
 mod node_config_loader;
+mod override_node_config;
 mod peer_monitoring_config;
 mod persistable_config;
 mod quorum_store_config;
@@ -46,6 +47,7 @@ pub use mempool_config::*;
 pub use network_config::*;
 pub use node_config::*;
 pub use node_config_loader::sanitize_node_config;
+pub use override_node_config::*;
 pub use peer_monitoring_config::*;
 pub use persistable_config::*;
 pub use quorum_store_config::*;
